lightningd/channel: derive local channel_id.
We weren't setting it, but the peer wasn't checking it either.
